+++ test/rack_suggest_test.dart
@@ -19,6 +19,7 @@ class FakePersistence extends GamePersistence {
     required players,
     required bag,
     required currentPlayerIndex,
+    Map<String, int> tileOwners = const {},
   }) async {}
   @override
   SavedGame? load() => null;